enhancements
Extended the plug in to capture Up/Down key presses which allows you to
move through the list without using a mouse, you can also hit enter to
select the currently selected item.

Included the scrollintoview js file which makes sure the currently
selected item in the list is viewable if its contained within a
scrolling element.

Changes to the CSS.